Move Coordinator enum to types file
