+++ model/src/conditions/index.ts
@@ -22,4 +22,5 @@ export { Condition } from "./condition";
 export { ConditionRef } from "./condition-ref";
 export { ConditionGroup } from "./condition-group";
 export { ConditionsModel } from "./inline-condition-model";
-export { Coordinator, toExpression, toPresentationString } from "./helpers";
+export { toExpression, toPresentationString } from "./helpers";
+export { Coordinator } from "./types";
